requirements.txt
python-ipmi==0.5.4
paho-mqtt==1.6.1